Split TalksDateMenu and it's item
