+++ src/feedback/talks/TalksDateMenu.js
@@ -5,12 +5,10 @@ import {
     getProjectIdSelector,
     getProjectSelectedDateSelector,
 } from '../project/projectSelectors'
-import { COLORS } from '../../constants/colors'
-import { Link } from 'react-router-dom'
-import { DateTime } from 'luxon'
 import makeStyles from '@material-ui/core/styles/makeStyles'
 import { useTranslation } from 'react-i18next'
 import { TALK_NO_DATE } from '../../core/talks/talksUtils'
+import { TalkDateMenuItem } from './TalkDateMenuItem'
 
 const useStyles = makeStyles({
     menu: {
@@ -34,7 +32,7 @@ const TalksDateMenu = () => {
     return (
         <div className={classes.menu}>
             {talksDates.map((date) => (
-                <Item
+                <TalkDateMenuItem
                     key={date}
                     date={date}
                     url={`/${currentProjectId}/${
@@ -44,55 +42,6 @@ const TalksDateMenu = () => {
                     isSelected={selectedDate === date}
                 />
             ))}
-        </div>
-    )
-}
-
-const useStylesItem = makeStyles((theme) => ({
-    item: {
-        color: (props) =>
-            props.isSelected
-                ? theme.palette.text.primary
-                : theme.palette.text.secondary,
-        borderBottom: (props) =>
-            props.isSelected ? `2px ${COLORS.RED_ORANGE} solid` : 'none',
-        '&:hover': {
-            color: COLORS.RED_ORANGE,
-        },
-    },
-    a: {
-        padding: 10,
-        color: 'inherit',
-        display: 'block',
-    },
-}))
-
-const Item = ({ date, url, isSelected, noDateLabel }) => {
-    const { t } = useTranslation()
-    const classes = useStylesItem({
-        isSelected,
-    })
-
-    const getDateLabel = (date) => {
-        if (date !== TALK_NO_DATE) {
-            return DateTime.fromISO(date).toLocaleString({
-                weekday: 'long',
-                day: 'numeric',
-            })
-        }
-        return noDateLabel
-    }
-
-    const label = getDateLabel(date)
-
-    return (
-        <div className={classes.item}>
-            <Link
-                to={`${url}`}
-                className={classes.a}
-                title={t('talks.date') + label}>
-                {label}
-            </Link>
         </div>
     )
 }
